Optimize day 7, part 2 somewhat
Better, but still prohibitively slow. Also, scratch that “it *might*
finish today” from the last commit – runtime might still work out if I
had infinite RAM, but memory usage is growing enough that it would run
out of that (and swap) if I let it run further. (I got as far as the 70s
out of a total of 140ish lines, with steadily increasing time and memory
usage per line of course.)