Optimize day 7, part 2 a bit more
Only keep the last line of each world in memory, rather than the whole
history. This still won’t be enough – we need to not be “materializing”
all these worlds at all, and avoid calculating the same line millions of
times – but it still makes it go noticeably faster because it has to do
a lot less memory management already.