Optimize a tiny bit
I was hoping for more of an improvement from this, but skipping the
circuit merging if the connection we’re looking at is fully inside one
circuit only reduced it to 2m54,910s, i.e. shaving off some 6 seconds.
Pity. (Also, note that returning early means we’re skipping the
assertions at the end, but I can live with that.)