Correcter optimization
This is, I think, the correct way to implement the optimization I tried
to add in df9b413b1a and had to revert in cae70413cb. If a point
*strictly inside the rectangle* is not aligned with any green lines,
then we don’t need to check if it’s inside or outside, because one of
its aligned neighbors ought to have the same result. (However, points on
the edge of the rectangle still need to be checked even if they’re not
aligned.)